Add shop scene. Add money and item picking functionalities in River.

+++ Assets/Scripts/ItemSystem/ShopSystem/ShopItemData.cs
@@ -5,7 +5,7 @@ namespace Shop {
 [CreateAssetMenu(fileName = "ItemData", menuName = "ReincarnationCompany/ShopItemData", order = 0)]
 public class ShopItemData : ScriptableObject {
     public BaseItemData baseData;
-    public float price;
+    public int price;
 }
 
 }
